requirements.txt
streamlit==1.32.2
opencv-python==4.9.0.80
numpy==1.26.4
ultralytics==8.1.27
onnxruntime==1.17.3
google-genai==0.3.0
python-dotenv==1.0.1
